Moved actual logic to `/src/bin/packet_handlers/` to be consistent with filing scheme present in rest of project.
