+++ src/bin/src/packet_handlers/mod.rs
@@ -1,3 +1,4 @@
+mod animations;
 mod handshake;
 mod login_process;
 mod tick_handler;
